html/browsers/browsing-the-web/navigating-across-documents/012-blocking-then-async-postmessage.html

<!doctype html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/browsing-the-web.html#navigating-across-documents">
<title>Link with onclick, which executes a handler containing a 100ms blocking operation, then iframe document.write, then returns true.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
 <a target="test" onclick="(function() {((e,n)=>{var a=e();while(e()<a+n);})(Date.now,100); document.querySelector('iframe').contentDocument.write('<script>parent.postMessage(&quot;write-post-message&quot;, &quot;*&quot;)</script>');reports.push('onclick');return true;})()" href="href.html">Test</a>

<script>
var test = async_test();
var reports = [];
var iframe = document.querySelector('iframe');

window.onmessage = test.step_func(function(e) {
  reports.push(e.data);

  // "href" is expected last, if this is not received last or
  // not received at all, the test must fail.
  if (e.data === "href") {
    assert_equals(reports[0], "click");
    assert_equals(reports[1], "onclick");
    // This comes after "onclick" because it's
    // sent via postMessage
    assert_equals(reports[2], "write-post-message");
    assert_equals(reports[3], "href");
    assert_equals(reports.length, 4);
    test.done();
  }
});

test.step(function() {
  reports.push("click");
  document.querySelector("a").click();
});
</script>

<!--
1. This test creates an iframe and an anchor with both an `onclick` and `href`.
2. The test "clicks" the anchor, triggering the `onclick` handler operation,
    which will document.write a script to the iframe which sends a message "write".
3. Navigate the iframe to the anchor's `href`
4. If the semantics are implemented correctly, reports will be
    ["click", "onclick", "write-post-message", "href"]

-->
